requirements.txt
uvicorn[standard]==0.30.6
python-dotenv==1.2.1
fastapi==0.115.6
python-multipart==0.0.20
sqlalchemy==2.0.44
langchain-chroma==1.0.0
langchain-openai==1.0.3
langchain-google-genai==3.0.3
PyMuPDF==1.26.6
docling==2.62.0
langchain-text-splitters==1.0.0
python-jose==3.5.0
bcrypt==5.0.0